+++ kroxylicious-integration-tests/src/test/java/io/kroxylicious/proxy/FilterIT.java
@@ -580,49 +580,4 @@ class FilterIT {
         }
     }
 
-    @Test
-    void shouldInitiate(@SaslMechanism(principals = { @SaslMechanism.Principal(user = "alice", password = "alice-secret") }) KafkaCluster cluster,
-                        Topic topic)
-            throws Exception {
-        String testName = "shouldInitiate";
-
-        NamedFilterDefinition saslInitiation = new NamedFilterDefinitionBuilder(
-                SaslPlainInitiation.class.getName(),
-                SaslPlainInitiation.class.getName())
-                .build();
-        NamedFilterDefinition lawyer = new NamedFilterDefinitionBuilder(
-                ClientTlsAwareLawyer.class.getName(),
-                ClientTlsAwareLawyer.class.getName())
-                .build();
-        var config = proxy(cluster)
-                .addToFilterDefinitions(saslInitiation, lawyer)
-                .addToDefaultFilters(saslInitiation.name(), lawyer.name());
-
-        try (var tester = kroxyliciousTester(config);
-                var producer = tester.producer(Map.of(
-                        CLIENT_ID_CONFIG, testName + "-producer",
-                        DELIVERY_TIMEOUT_MS_CONFIG, 3_600_000));
-                var consumer = tester
-                        .consumer(Serdes.String(), Serdes.ByteArray(), Map.of(
-                                CLIENT_ID_CONFIG, testName + "-consumer",
-                                GROUP_ID_CONFIG, "my-group-id",
-                                AUTO_OFFSET_RESET_CONFIG, "earliest"))) {
-            producer.send(new ProducerRecord<>(topic.name(), "my-key", PLAINTEXT)).get();
-
-            producer.flush();
-
-            consumer.subscribe(Set.of(topic.name()));
-            var records = consumer.poll(Duration.ofSeconds(10));
-
-            assertThat(records).hasSize(1);
-            assertThat(records.records(topic.name()))
-                    .as("topic %s records", topic.name())
-                    .singleElement()
-                    .extracting(ConsumerRecord::value)
-                    .as("record value")
-                    .extracting(String::new)
-                    .isEqualTo(PLAINTEXT);
-        }
-    }
-
 }
